hide some items from context menu if user can not write

+++ apps/studio/src/components/sidebar/ConnectionSidebar.vue
@@ -570,6 +570,7 @@ export default {
         : null
       this.$modal.show('connection-folder-modal')
     },
+    /** @param folder {import("@/common/interfaces/ISavedQuery").default} */
     showFolderContextMenu(event, folder) {
       if (event.target.tagName === 'INPUT') {
         return;
@@ -577,19 +578,25 @@ export default {
       event.stopPropagation();
 
       const options = []
+      const canWrite = folder.canWrite ?? true;
       if (this.isCloud && !folder.parentId) {
         options.push({ name: 'New Subfolder', handler: ({ item }) => this.createSubfolder(item) })
       }
-      if (folder.parentId) {
+      // =========== "Move to ..." options =========
+      if (!canWrite) {
+        // do nothing
+      } else if (folder.parentId) {
         const otherRoots = this.rootFolders.filter(f => f.id !== folder.parentId)
         otherRoots.forEach(root => {
           options.push({ name: `Move to ${root.name}`, handler: ({ item }) => this.moveFolderToParent(item, root) })
         })
       }
-      options.push(
-        { name: 'Rename', handler: ({ item }) => this.renameFolder(item) },
-        { name: 'Delete', handler: ({ item }) => this.deleteFolder(item) }
-      )
+      if (canWrite) {
+        options.push(
+          { name: 'Rename', handler: ({ item }) => this.renameFolder(item) },
+          { name: 'Delete', handler: ({ item }) => this.deleteFolder(item) }
+        )
+      }
       this.$bks.openMenu({ event, item: folder, options })
     },
     createSubfolder(parentFolder) {
